+++ src/main/java/cz/pfreiberg/knparser/domain/prvkykatastralnimapy/ObrazyBoduBp.java
@@ -20,7 +20,7 @@ public class ObrazyBoduBp {
 	private Integer priznakKontextu;
 	private Long rizeniIdVzniku;
 	private Long rizeniIdZaniku;
-	private Integer typppdKod;
+	private Long typppdKod;
 	private Double souradniceY;
 	private Double souradniceX;
 	private String text;
@@ -86,11 +86,11 @@ public class ObrazyBoduBp {
 		this.rizeniIdZaniku = rizeniIdZaniku;
 	}
 
-	public Integer getTypppdKod() {
+	public Long getTypppdKod() {
 		return typppdKod;
 	}
 
-	public void setTypppdKod(Integer typppdKod) {
+	public void setTypppdKod(Long typppdKod) {
 		this.typppdKod = typppdKod;
 	}
