+++ src/main/java/cz/pfreiberg/knparser/domain/prvkykatastralnimapy/ObrazyBudov.java
@@ -20,7 +20,7 @@ public class ObrazyBudov {
 	private Integer priznakKontextu;
 	private Long rizeniIdVzniku;
 	private Long rizeniIdZaniku;
-	private Integer typppdKod;
+	private Long typppdKod;
 	private Double souradniceY;
 	private Double souradniceX;
 	private Double velikost;
@@ -84,11 +84,11 @@ public class ObrazyBudov {
 		this.rizeniIdZaniku = rizeniIdZaniku;
 	}
 
-	public Integer getTypppdKod() {
+	public Long getTypppdKod() {
 		return typppdKod;
 	}
 
-	public void setTypppdKod(Integer typppdKod) {
+	public void setTypppdKod(Long typppdKod) {
 		this.typppdKod = typppdKod;
 	}
